Added my first Ruby assignment.
assignment12-1: Takes the amount a user inputs and calculates the 10%, 15%, and 20% tip.

assignment12-2: Takes an integer of how many scores the user would like to adverage then allows the user to input all the scores and it will adverage the scores.

assignment12-3: Allows the user to input a base and then calculates the power of that base until the power of the base and a number is > 1,000,000.